+++ example/lib/main.dart
@@ -11,7 +11,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Smooth Page Indcator Demo',
+      title: 'Smooth Page Indicator Demo',
       theme: ThemeData(primarySwatch: Colors.blue),
       home: HomePage(),
     );
@@ -24,12 +24,28 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  final controller = PageController(viewportFraction: 0.8);
+  final controller = PageController(viewportFraction: 0.8, keepPage: true);
 
   @override
   Widget build(BuildContext context) {
+    final pages = List.generate(
+        6,
+        (index) => Container(
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.circular(16),
+                color: Colors.grey.shade300,
+              ),
+              margin: EdgeInsets.symmetric(horizontal: 10, vertical: 4),
+              child: Container(
+                height: 280,
+                child: Center(
+                    child: Text(
+                  "Page $index",
+                  style: TextStyle(color: Colors.indigo),
+                )),
+              ),
+            ));
     return Scaffold(
-      backgroundColor: Colors.grey.shade300,
       body: SafeArea(
         child: SingleChildScrollView(
           child: Column(
@@ -37,99 +53,134 @@ class _HomePageState extends State<HomePage> {
             children: <Widget>[
               SizedBox(height: 16),
               SizedBox(
-                height: 300,
-                child: PageView(
+                height: 240,
+                child: PageView.builder(
                   controller: controller,
-                  children: List.generate(
-                      6,
-                      (_) => Card(
-                            shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(16)),
-                            margin: EdgeInsets.symmetric(
-                                horizontal: 10, vertical: 4),
-                            child: Container(height: 280),
-                          )),
-                ),
-              ),
-              SizedBox(height: 16),
-              Padding(
-                padding: const EdgeInsets.only(top: 16, bottom: 8),
-                child: Text('Worm'),
-              ),
-              Container(
-                child: SmoothPageIndicator(
-                  controller: controller,
-                  count: 6,
-                  effect: WormEffect(),
+                  // itemCount: pages.length,
+                  itemBuilder: (_, index) {
+                    return pages[index % pages.length];
+                  },
                 ),
               ),
               Padding(
-                padding: const EdgeInsets.only(top: 16, bottom: 8),
-                child: Text('Expanding Dots '),
-              ),
-              Container(
-                child: SmoothPageIndicator(
-                  controller: controller,
-                  count: 6,
-                  effect: ExpandingDotsEffect(
-                    expansionFactor: 4,
-                  ),
+                padding: const EdgeInsets.only(top: 24, bottom: 12),
+                child: Text(
+                  'Worm',
+                  style: TextStyle(color: Colors.black54),
                 ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8),
-                child: Text('Jumping Dot '),
               ),
               SmoothPageIndicator(
                 controller: controller,
-                count: 6,
-                effect: JumpingDotEffect(),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8),
-                child: Text('slide Dots '),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top: 16, bottom: 8),
-                child: Text('Scrolling Dots '),
-              ),
-              SmoothPageIndicator(
-                  controller: controller,
-                  count: 6,
-                  effect: ScrollingDotsEffect(
-                    activeStrokeWidth: 2.6,
-                    activeDotScale: .4,
-                    radius: 8,
-                    spacing: 10,
-                  )),
-              Padding(
-                padding: const EdgeInsets.only(top: 16, bottom: 8),
-                child: Text('Scale'),
-              ),
-              Container(
-                child: SmoothPageIndicator(
-                  controller: controller,
-                  count: 6,
-                  effect: ScaleEffect(),
+                count: pages.length,
+                effect: WormEffect(
+                  dotHeight: 16,
+                  dotWidth: 16,
+                  type: WormType.thin,
+                  // strokeWidth: 5,
                 ),
               ),
               Padding(
                 padding: const EdgeInsets.only(top: 16, bottom: 8),
-                child: Text('Slide'),
+                child: Text(
+                  'Jumping Dot',
+                  style: TextStyle(color: Colors.black54),
+                ),
               ),
               Container(
                 child: SmoothPageIndicator(
                   controller: controller,
-                  count: 6,
-                  effect: SlideEffect(
-                    spacing: 8.0,
-                    radius: 4.0,
-                    dotWidth: 24.0,
-                    dotHeight: 16.0,
-                    dotColor: Colors.grey,
-                    paintStyle: PaintingStyle.stroke,
-                    strokeWidth: 2,
-                    activeDotColor: Colors.indigo,
+                  count: pages.length,
+                  effect: JumpingDotEffect(
+                    dotHeight: 16,
+                    dotWidth: 16,
+                    jumpScale: .7,
+                    verticalOffset: 15,
+                  ),
+                ),
+              ),
+              // Padding(
+              //   padding: const EdgeInsets.only(top: 16, bottom: 12),
+              //   child: Text(
+              //     'Swap',
+              //     style: TextStyle(color: Colors.black54),
+              //   ),
+              // ),
+              // Container(
+              //   child: SmoothPageIndicator(
+              //     controller: controller,
+              //     count: pages.length,
+              //     effect: SwapEffect(
+              //       dotHeight: 16,
+              //       dotWidth: 16,
+              //       type: SwapType.yRotation,
+              //     ),
+              //   ),
+              // ),
+              Padding(
+                padding: const EdgeInsets.only(top: 16, bottom: 12),
+                child: Text(
+                  'Scrolling Dots',
+                  style: TextStyle(color: Colors.black54),
+                ),
+              ),
+              SmoothPageIndicator(
+                  controller: controller,
+                  count: pages.length,
+                  effect: ScrollingDotsEffect(
+                    activeStrokeWidth: 2.6,
+                    activeDotScale: 1.3,
+                    maxVisibleDots: 5,
+                    radius: 8,
+                    spacing: 10,
+                    dotHeight: 12,
+                    dotWidth: 12,
+                  )),
+              Padding(
+                padding: const EdgeInsets.only(top: 16, bottom: 16),
+                child: Text(
+                  'Customizable Effect',
+                  style: TextStyle(color: Colors.black54),
+                ),
+              ),
+              Container(
+                // color: Colors.red.withOpacity(.4),
+                child: SmoothPageIndicator(
+                  controller: controller,
+                  count: pages.length,
+                  effect: CustomizableEffect(
+                    activeDotDecoration: DotDecoration(
+                      width: 32,
+                      height: 12,
+                      color: Colors.indigo,
+                      rotationAngle: 180,
+                      verticalOffset: -10,
+                      borderRadius: BorderRadius.circular(24),
+                      // dotBorder: DotBorder(
+                      //   padding: 2,
+                      //   width: 2,
+                      //   color: Colors.indigo,
+                      // ),
+                    ),
+                    dotDecoration: DotDecoration(
+                      width: 24,
+                      height: 12,
+                      color: Colors.grey,
+                      // dotBorder: DotBorder(
+                      //   padding: 2,
+                      //   width: 2,
+                      //   color: Colors.grey,
+                      // ),
+                      // borderRadius: BorderRadius.only(
+                      //     topLeft: Radius.circular(2),
+                      //     topRight: Radius.circular(16),
+                      //     bottomLeft: Radius.circular(16),
+                      //     bottomRight: Radius.circular(2)),
+                      borderRadius: BorderRadius.circular(16),
+                      verticalOffset: 0,
+                    ),
+                    spacing: 6.0,
+                    // activeColorOverride: (i) => colors[i],
+                    inActiveColorOverride: (i) => colors[i],
                   ),
                 ),
               ),
@@ -141,3 +192,12 @@ class _HomePageState extends State<HomePage> {
     );
   }
 }
+
+final colors = const [
+  Colors.red,
+  Colors.green,
+  Colors.greenAccent,
+  Colors.amberAccent,
+  Colors.blue,
+  Colors.amber,
+];