+++ lib/src/effects/jumping_dot_effect.dart
@@ -1,16 +1,22 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:smooth_page_indicator/src/painters/indicator_painter.dart';
 import 'package:smooth_page_indicator/src/painters/jumping_dot_painter.dart';
 
 import 'indicator_effect.dart';
 
-class JumpingDotEffect extends IndicatorEffect {
-  // Defines how high the dot will jump
-  final double elevation;
+class JumpingDotEffect extends BasicIndicatorEffect {
+  // The maximum scale the dot will hit while jumping
+  final double jumpScale;
+
+  // The vertical offset of the jumping dot
+  final double verticalOffset;
 
   const JumpingDotEffect({
     Color activeDotColor = Colors.indigo,
-    this.elevation = 15.0,
+    this.jumpScale = 1.4,
+    this.verticalOffset = 0.0,
     double offset = 16.0,
     double dotWidth = 16.0,
     double dotHeight = 16.0,
@@ -28,6 +34,14 @@ class JumpingDotEffect extends IndicatorEffect {
             paintStyle: paintStyle,
             dotColor: dotColor,
             activeDotColor: activeDotColor);
+
+  @override
+  Size calculateSize(int count) {
+    return Size(
+      dotWidth * count + (spacing * (count - 1)),
+      max(dotHeight, dotHeight * jumpScale) + verticalOffset.abs(),
+    );
+  }
 
   @override
   IndicatorPainter buildPainter(int count, double offset) {