+++ lib/src/painters/scale_painter.dart
@@ -3,7 +3,7 @@ import 'package:smooth_page_indicator/src/effects/scale_effect.dart';
 
 import 'indicator_painter.dart';
 
-class ScalePainter extends IndicatorPainter {
+class ScalePainter extends BasicIndicatorPainter {
   final ScaleEffect effect;
 
   ScalePainter({
@@ -22,23 +22,21 @@ class ScalePainter extends IndicatorPainter {
 
     var dotOffset = offset - current;
     var activeScale = effect.scale - 1.0;
-
     for (var index = 0; index < count; index++) {
-      canvas.drawRRect(_calcBounds(size.height, index), dotPaint);
+      var dot = _calcBounds(size.height, index);
+      canvas.drawRRect(dot.inflate(effect.activeStrokeWidth / 2), dotPaint);
       var color = effect.dotColor;
       var scale = 0.0;
       if (index == current) {
         scale = (effect.scale) - (activeScale * dotOffset);
         // ! Both a and b are non nullable
         color = Color.lerp(effect.activeDotColor, effect.dotColor, dotOffset)!;
-      } else if (index - 1 == current) {
+      } else if (index - 1 == current || (index == 0 && offset > count - 1)) {
         scale = 1.0 + (activeScale * dotOffset);
         // ! Both a and b are non nullable
-        color = Color.lerp(
-            effect.activeDotColor, effect.dotColor, 1.0 - dotOffset)!;
+        color = Color.lerp(effect.dotColor, effect.activeDotColor, dotOffset)!;
       }
-      canvas.drawRRect(
-          _calcBounds(size.height, index, scale), activePaint..color = color);
+      canvas.drawRRect(_calcBounds(size.height, index, scale), activePaint..color = color);
     }
   }
 
@@ -46,11 +44,8 @@ class ScalePainter extends IndicatorPainter {
     var width = effect.dotWidth * scale;
     var height = effect.dotHeight * scale;
     var startingPoint = effect.dotWidth + effect.spacing / 2;
-    var xPos = startingPoint / 2 -
-        width / 2 +
-        (offset * (effect.dotWidth + effect.spacing));
+    var xPos = startingPoint / 2 - width / 2 + (offset * (effect.dotWidth + effect.spacing));
     var yPos = canvasHeight / 2;
-    return RRect.fromLTRBR(xPos, yPos - height / 2, xPos + width,
-        yPos + height / 2, dotRadius * scale);
+    return RRect.fromLTRBR(xPos, yPos - height / 2, xPos + width, yPos + height / 2, dotRadius * scale);
   }
 }